requirements.txt
aiogram==3.13.1
asyncpg
sqlalchemy[asyncio]==2.0.36
alembic==1.13.2
python-dotenv==1.0.1
aiofiles==23.2.1
aiosmtplib==3.0.2
email-validator==2.1.1
bcrypt==4.1.3
python-dateutil==2.9.0.post0
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4

